# perf: Ensure cluster topology building completes before giving up RLock